Move ci-op recursive config loading from pkg/load/agents to pkg/load
Currently, the code to move all ci-op configs is part of the agent.
However there are cases where this is needed but an agent is not, e.G. a
CLI that acts on all configs.
